migrate getSvg code --> reusable .js
