+++ web/src/js/get_svg.js
@@ -1,7 +1,186 @@
-// based on https://github.com/mbostock/svjimmy/blob/master/index.js
-//
+const prefix = {
+  xmlns: "http://www.w3.org/2000/xmlns/",
+  xlink: "http://www.w3.org/1999/xlink",
+  svg: "http://www.w3.org/2000/svg"
+}
 
-export function getPng(selector, download = false, filename="outbreakinfo_visualization.png") {
+// code adapted from https://github.com/nytimes/svg-crowbar (thanks, Mike Bostock)
+export function getSvg(figureRef) {
+  const refs = document.getElementsByClassName(figureRef);
+  console.log(refs)
+  var emptySvg = window.document.createElementNS(prefix.svg, 'svg');
+  window.document.body.appendChild(emptySvg);
+  var emptySvgDeclarationComputed = getComputedStyle(emptySvg);
+
+  const svgObject = getSvgSources(refs, emptySvgDeclarationComputed);
+  return (svgObject)
+}
+
+function getSvgSources(svgs, emptySvgDeclarationComputed) {
+  var svgInfo = [];
+  var doctype = '<?xml version="1.0" standalone="no"?><!DOCTYPE svg PUBLIC "-//W3C//DTD SVG 1.1//EN" "http://www.w3.org/Graphics/SVG/1.1/DTD/svg11.dtd">';
+
+  // apparently nodes, while array-like, don't have a `forEach` property attached to them... hence this syntax
+  [].forEach.call(svgs, svg => {
+
+    svg.setAttribute("version", "1.1");
+
+    // removing attributes so they aren't doubled up
+    svg.removeAttribute("xmlns");
+    svg.removeAttribute("xlink");
+
+
+    // These are needed for the svg
+    if (!svg.hasAttributeNS(prefix.xmlns, "xmlns")) {
+      svg.setAttributeNS(prefix.xmlns, "xmlns", prefix.svg);
+    }
+
+    if (!svg.hasAttributeNS(prefix.xmlns, "xmlns:xlink")) {
+      svg.setAttributeNS(prefix.xmlns, "xmlns:xlink", prefix.xlink);
+    }
+
+    // necessary to nest styles inline
+    setInlineStyles(svg, emptySvgDeclarationComputed);
+
+    var source = (new XMLSerializer()).serializeToString(svg);
+    var rect = svg.getBoundingClientRect();
+
+    const title = svg.getAttribute("name");
+    const footer = getFooter(rect.width, rect.height);
+    const header = getHeader(rect.width, title);
+
+
+    svgInfo.push({
+      top: rect.top,
+      left: rect.left,
+      width: rect.width,
+      height: rect.height,
+      class: svg.getAttribute("class"),
+      id: svg.getAttribute("id"),
+      name: title,
+      childElementCount: svg.childElementCount,
+      source: [doctype + `<svg width="${rect.width}" height="${rect.height + 100}">` + header + '<g transform="translate(0,35)">' + source + footer + "</g></svg>"]
+    });
+  });
+  return svgInfo;
+}
+
+function getHeader(width, title) {
+  return (`<svg id="title" width="${width}" height="20">
+  <text x="0" y="0" transform="translate(10,10)" fill="currentColor"
+    style="dominant-baseline: hanging; font-size:18px;display:block;font-family:&quot;DM Sans&quot;, Avenir, Helvetica, Arial, sans-serif;height:auto;line-height:15px;outline-color:rgb(44, 62, 80);overflow-x:visible;overflow-y:visible;text-align:center;text-decoration:none solid rgb(44, 62, 80);text-decoration-color:rgb(44, 62, 80);vertical-align:baseline;white-space:nowrap;width:auto;column-rule-color:rgb(44, 62, 80);-webkit-font-smoothing:antialiased;perspective-origin:0px 0px;-webkit-text-emphasis-color:rgb(44, 62, 80);-webkit-text-fill-color:rgb(44, 62, 80);-webkit-text-stroke-color:rgb(44, 62, 80);transform-origin:0px 0px;fill:rgb(44, 62, 80);text-anchor:start;caret-color:rgb(44, 62, 80);">
+  ${title}</text>
+  </svg>`)
+}
+
+function getFooter(width, height) {
+  return (`<svg width="${width}" height="50" id="footer" class="sources mt-2" transform="translate(0, ${height + 15})">
+  <g id="background">
+  <rect width="${width}" height="50" style="fill: #dee2e6"></rect>
+  </g>
+
+      <g id="top_border">
+        <line x1="0" y1="2" x2="${width}" y2="2" stroke="#126B93" stroke-width="5"></line>
+      </g>
+
+  <g id="citation" transform="translate(10,5)">
+    <g transform="translate(0, -8)" id="logo">
+      <svg version="1.1" id="Layer_1" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" x="0px" y="0px" width="25px"
+         viewBox="0 0 396.4 396.4" style="enable-background:new 0 0 396.4 396.4;" xml:space="preserve">
+      <g>
+        <circle cx="198.2" cy="203" r="180"/>
+        <g>
+          <circle style="fill:#114068;" cx="198.2" cy="204.7" r="151.6"/>
+        </g>
+        <linearGradient id="SVGID_1_" gradientUnits="userSpaceOnUse" x1="195.2239" y1="53.7053" x2="333.4954" y2="293.1985">
+          <stop  offset="0" style="stop-color:#39C2E2"/>
+          <stop  offset="0.6781" style="stop-color:#39C2E2;stop-opacity:0"/>
+        </linearGradient>
+        <path style="fill:url(#SVGID_1_);" d="M196.1,212.6l114.2,94c24.5-26.9,39.6-62.6,39.6-101.9c0-83.7-67.9-151.6-151.6-151.6
+          c-0.7,0-1.4,0.1-2.1,0.1V212.6z"/>
+        <g>
+          <path style="fill:#FFFFFF;" d="M198.2,272.2c-32.4,0-58.8-26.4-58.8-58.8s26.4-58.8,58.8-58.8s58.8,26.4,58.8,58.8
+            S230.6,272.2,198.2,272.2z M198.2,161.6c-28.5,0-51.7,23.2-51.7,51.7s23.2,51.7,51.7,51.7c28.5,0,51.7-23.2,51.7-51.7
+            S226.7,161.6,198.2,161.6z"/>
+        </g>
+        <g>
+          <path style="fill:#FFFFFF;" d="M198.2,330.8c-64.7,0-117.4-52.7-117.4-117.4S133.5,95.9,198.2,95.9c64.7,0,117.4,52.7,117.4,117.4
+            S262.9,330.8,198.2,330.8z M198.2,103c-60.8,0-110.3,49.5-110.3,110.3s49.5,110.3,110.3,110.3s110.3-49.5,110.3-110.3
+            S259,103,198.2,103z"/>
+        </g>
+        <g>
+          <circle style="fill:#D13B62;" cx="269.7" cy="172.8" r="18.7"/>
+        </g>
+        <g>
+          <circle style="fill:#D13B62;" cx="126.5" cy="259.3" r="9.4"/>
+        </g>
+        <g>
+          <circle style="fill:#D13B62;" cx="225.3" cy="259.3" r="27.1"/>
+        </g>
+        <path style="fill:#FFFFFF;" d="M194.6,23c-0.3,0.6,0,2.4,0,3.1v186.4c0,2.3,1.6,4.1,3.6,4.1s3.6-1.9,3.6-4.1V26.1
+          c0-0.7,0.3-2.5,0-3.1c-0.9,0-2,0-2.9,0C197.6,23,195.8,23,194.6,23z"/>
+      </g>
+      </svg>
+      </g>
+      <g id="outbreak-info" transform="translate(0, 22.3)">
+        <text x="0" y="0" transform="translate(30,0)" style="font-size:17px; dominant-baseline: middle;font-family:&quot;DM Sans&quot;, Avenir, Helvetica, Arial, sans-serif;">outbreak.info</text>
+        <text x="0" y="0" transform="translate(${width - 30},0)" style="dominant-baseline: middle; font-size:13px; text-anchor: end;font-family:&quot;DM Sans&quot;, Avenir, Helvetica, Arial, sans-serif;">${'formatDate(this.today, "%d %B %Y")'}</text>
+      </g>
+
+      <g id="sources" transform="translate(0, 36)">
+        <text x="0" y="0" transform="translate(30,0)" style="dominant-baseline: middle;font-size: 10px;fill: #6c757d;font-family:&quot;DM Sans&quot;, Avenir, Helvetica, Arial, sans-serif;">Source: ${"this.sourceString"}</text>
+      </g>
+      </g>
+    </svg>`)
+}
+
+function setInlineStyles(svg, emptySvgDeclarationComputed) {
+
+  function explicitlySetStyle(element) {
+    var cSSStyleDeclarationComputed = getComputedStyle(element);
+    var i, len, key, value;
+    var computedStyleStr = "";
+    for (i = 0, len = cSSStyleDeclarationComputed.length; i < len; i++) {
+      key = cSSStyleDeclarationComputed[i];
+      value = cSSStyleDeclarationComputed.getPropertyValue(key);
+      if (value !== emptySvgDeclarationComputed.getPropertyValue(key)) {
+        computedStyleStr += key + ":" + value + ";";
+      }
+    }
+    element.setAttribute('style', computedStyleStr);
+  }
+
+  function traverse(obj) {
+    var tree = [];
+    tree.push(obj);
+    visit(obj);
+
+    function visit(node) {
+      if (node && node.hasChildNodes()) {
+        var child = node.firstChild;
+        while (child) {
+          if (child.nodeType === 1 && child.nodeName != 'SCRIPT') {
+            tree.push(child);
+            visit(child);
+          }
+          child = child.nextSibling;
+        }
+      }
+    }
+    return tree;
+  }
+  // hardcode computed css styles inside svg
+  var allElements = traverse(svg);
+  var i = allElements.length;
+  while (i--) {
+    explicitlySetStyle(allElements[i]);
+  }
+}
+
+// based on https://github.com/mbostock/svjimmy/blob/master/index.js
+// Thanks, Mike.
+
+export function getPng(selector, download = false, filename = "outbreakinfo_visualization.png") {
   return new Promise((resolve, reject) => {
     // const refs = document.getElementsByClassName(classID);
     var document = global.document,
@@ -77,6 +256,5 @@ export function getPng(selector, download = false, filename="outbreakinfo_visual
       canvas.height = height;
       image.src = imageUrl;
     });
-    // reject("ERROR: no svg found with that selector")
   })
 }
